chore(swiss-ai-hub): Move ThreadFilters into entities/types package
