+++ packages/core/swiss_ai_hub/core/persistence/messaging/entities/tests/unit/test_thread_filters.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from swiss_ai_hub.core.persistence.messaging.entities.thread_filters import ThreadFilters
+from swiss_ai_hub.core.persistence.messaging.entities.types.thread_filters import ThreadFilters
 
 
 class TestThreadFilters:
