+++ packages/core/swiss_ai_hub/core/persistence/messaging/entities/types/thread_filters.py
